error refactor
