--- locales/en-us/pageNotFound.json
@@ -1,6 +1,0 @@
-{
-	"title": "Page not found",
-	"text": "The page you are trying to access does not exist.",
-	"home": "Go to home page",
-	"errorOcurred": "An error occurred (✖╭╮✖)"
-}
