--- locales/es-ar/pageNotFound.json
@@ -1,6 +1,0 @@
-{
-	"title": "Página no encontrada",
-	"text": "La página a la que intentás acceder no existe.",
-	"home": "Ir al inicio",
-	"errorOcurred": "Ocurrió un error (✖╭╮✖)"
-}
