--- locales/pt-br/pageNotFound.json
@@ -1,6 +1,0 @@
-{
-	"title": "Página não encontrada",
-	"text": "A página que você está tentando acessar não existe.",
-	"home": "Vá para iniciar",
-	"errorOcurred": "Um erro ocorreu (✖╭╮✖)"	
-}
