+++ locales/pt-br/pbError.json
@@ -1,0 +1,8 @@
+{
+	"home": "Vá para iniciar",
+	"errorOcurred": "Um erro ocorreu (✖╭╮✖)",
+	"pageNotFound": {
+		"title": "Página não encontrada",
+		"text": "A página que você está tentando acessar não existe."
+	}
+}
