+++ src/App.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import './App.css';
 import { createHashRouter, RouterProvider, Outlet } from "react-router-dom";
 import { Home } from './components/home/Home';
-import { PBError } from './components/PBError';
+import { PBError } from './components/pageNotFound/PBError';
 import { ChallengeById, ChallengeByName } from './components/ChallengeView';
 import { BookView } from './components/book/BookView';
 import { ImportedChallengeView } from './components/ImportedChallengeView';
